+++ src/main/java/com/example/policy/service/AllOfService.java
@@ -1,17 +1,27 @@
 package com.example.policy.service;
 
+import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import com.example.policy.model.AllOf;
-import com.example.policy.model.AllOf;
+import com.example.policy.model.AnyOf;
+import com.example.policy.model.Match;
 import com.example.policy.repository.AllOfRepository;
+import com.example.policy.repository.MatchRepository;
 
+import jakarta.persistence.EntityNotFoundException;
+
+@Service
 public class AllOfService {
 	@Autowired
 	private AllOfRepository allOfRepository;
+	
+	@Autowired
+	private MatchRepository matchRepository;
 	
 	public List<AllOf> getAllAllOfs(){
 		return allOfRepository.findAll();
@@ -31,6 +41,21 @@ public class AllOfService {
 	                });
 	}
 	public void deleteAllOf(Long id) {
-		 allOfRepository.deleteById(id);
+		 Optional<AllOf> allof=allOfRepository.findById(id);
+		 if(allof.isPresent()) {
+			List<Match> matchList = allof.get().getMatches();
+			Iterator<Match> iterator = matchList.iterator();
+	        while (iterator.hasNext()) {
+	        	Match match = iterator.next();
+	            iterator.remove();  // 從列表中移除 Match
+	            matchRepository.delete(match);  // 刪除 Match
+	        }
+	        AnyOf anyof = allof.get().getAnyOf();
+	        anyof.getAllOfs().remove(allof.get());
+			allOfRepository.deleteById(id);
+		 }else {
+			 throw new EntityNotFoundException("AllOf not found with id: " + id);
+		 }
+		 
 	}
 }
